refactor
Start off by downloading all maps before delegating to a render-map
directive to trigger the drawing via the render-layer directive.
Resulted in the removal of the render-neighborhoods directive as that
is now handled in the generic render layer. Of course, for this to all
work the heavy lifting of downloading and triggering the render is
part of a maps service.